+++ src/C2AllTestSuiteInfo.ts
@@ -18,7 +18,7 @@ import {TaskPool} from './TaskPool';
 export class C2AllTestSuiteInfo implements TestSuiteInfo, vscode.Disposable {
   readonly type: 'suite' = 'suite';
   readonly id: string;
-  readonly label: string = 'Catch2 Test Explorer (matepek)';
+  readonly label: string;
   readonly children: C2TestSuiteInfo[] = [];
   private readonly _executables: C2ExecutableInfo[] = [];
   private _isDisposed = false;
@@ -40,6 +40,7 @@ export class C2AllTestSuiteInfo implements TestSuiteInfo, vscode.Disposable {
       public execWatchTimeout: number,
       public isNoThrow: boolean,
   ) {
+    this.label = workspaceFolder.name + ' - Catch2';
     this.id = generateUniqueId();
   }