+++ src/test/FsWrapper.test.ts
@@ -10,11 +10,6 @@ import * as vscode from 'vscode';
 
 import {spawnAsync, statAsync} from '../FsWrapper';
 
-assert.notEqual(vscode.workspace.workspaceFolders, undefined);
-assert.equal(vscode.workspace.workspaceFolders!.length, 1);
-
-const workspaceFolderUri = vscode.workspace.workspaceFolders![0].uri;
-
 describe('FsWrapper.spawnAsync', function() {
   it('echoes', async function() {
     const isWin = process.platform === 'win32';
@@ -65,8 +60,7 @@ describe('FsWrapper.statAsync', function() {
   })
 
   it('exists', async function() {
-    const res = await statAsync(
-        path.join(workspaceFolderUri.fsPath, 'FsWrapper.test.js'));
+    const res = await statAsync(__filename);
     assert.ok(res.isFile());
     assert.ok(!res.isDirectory());
   })
@@ -131,6 +125,6 @@ describe('path', function() {
 
 describe('vscode.Uri', function() {
   it('!=', function() {
-    assert.ok(vscode.Uri.file(workspaceFolderUri.path) != workspaceFolderUri);
+    assert.ok(vscode.Uri.file(__filename) != vscode.Uri.file(__filename));
   })
 })